Clean up BaseModule, readd dataclass example
